fun_center_image {sunxplrr}	R Documentation

Centers the provided image

Description

Centers the provided image by shifting the x-axis and the y-axis of the image appropriately. Do not center sdo images, since they are centered already. The algorithm is slightly adapted from Berry, R. and Burnell, J.: The Handbook of Astronomical Image Processing, 2nd edition 2005, p. 323.

Usage


fun_center_image(image, hdrlst, header, sdo.image = "FALSE")

Arguments

image
tibble image with i and j pixel coordinates and pixel values x. Additional columns are ignored and will be not returned.

hdrlst
list containing image FITS header keywords and values.

sdo.image
boolean switch for dummy use in the case of non sdo calcium images.

Value

tibble with centered image.

Author(s)

[Thomas K. Friedli](mailto:thomas.friedli@bluewin.ch)

[Package sunxplrr version 0.1.0 Index]