fun_clv_calibrate {sunxplrr}	R Documentation

Calibrate the image

Description

Calibrate the image by adding the rescaled fitted degree-5 clv function to the residuals from the multi-parameter fit of the image.

Usage


fun_clv_calibrate(x, sdo.image = "FALSE")

Arguments

x
tibble containing columns with pixel coordinates i and j, pixel values x from the original image, fill from the disc mask, sclv and residuals from the sclv image.

sdo.image
boolean switch for dummy use in the case of non sdo calcium images.

Value

tibble containing the calibrated image values.

Author(s)

[Thomas K. Friedli](mailto:thomas.friedli@bluewin.ch)

[Package sunxplrr version 0.1.0 Index]