fun_clv_correct {sunxplrr}	R Documentation

Flattens an image

Description

Flattens a solar image according to a provided clv function.

Usage


fun_clv_correct(x, name.image = "calib", name.clv = "sclv")

Arguments

x
tibble containing columns with pixel coordinates i and j and columns with the pixel values of the image and the clv function.

name.image
name of the column containing the image pixel values.

name.clv
name of the column containing the clv function values.

Value

tibble with additional column to x containing the flattened image.

Author(s)

[Thomas K. Friedli](mailto:thomas.friedli@bluewin.ch)

[Package sunxplrr version 0.1.0 Index]