fun_clv_function {sunxplrr}	R Documentation

Calculates the standardized center-to-limb-variation function

Description

Calculates the center-to-limb-variation function according to formulae published by different authors, including Allen in Cox (2000), Pierce & Slaughter (1977) and Neckel & Labs (1994). Furthermore, a customly fitted clv may be calculated. Note, that this is the only available option for SDO images, since near the temperature minimum, a near flat clv is expected but the SDO images show a clear limb darkening.

Usage


fun_clv_function(x, clv.i0, sclv.method = "NL", sdo.image = "FALSE",
  clv.coeff = NULL)

Arguments

x
tibble containing columns with pixel coordinates i and j, pixel values x, pixel coordinates xi, yj and angle theta from an image disc.

clv.i0
central intensity of the fitted clv function of the image.

sdo.image
boolean switch for dummy use in the case of non sdo calcium images.

clv.coeff
vector with 6 coefficients from the fitted clv function.

method
implemented are "Allen", "PS", "NL" and "fit".

Value

z tibble with additional column to x containing the standardized clv function and the cf vector with coefficients of the standardized clv function.

Author(s)

[Thomas K. Friedli](mailto:thomas.friedli@bluewin.ch)

[Package sunxplrr version 0.1.0 Index]