fun_clv_rmplane {sunxplrr}	R Documentation

Subtracts plane from image

Description

Subtracts a previously fitted plane from the provided image values.

Usage


fun_clv_rmplane(x, hdrlst, coeff)

Arguments

x
tibble containing columns with pixel coordinates i and j, pixel values x from the original image, fill from the disc mask, image from the extracted image disc and theta.

hdrlst
list containing image FITS header keywords and values.

coeff
containing the previously fitted coefficients of an eight-parameter fit function.

Value

tibble containing the corrected image values.

Author(s)

[Thomas K. Friedli](mailto:thomas.friedli@bluewin.ch)

[Package sunxplrr version 0.1.0 Index]