fun_date2jd {sunxplrr}	R Documentation

Calculates Julian Date from a vector containing date with time objects

Description

Calculates Julian Date from a vector containing date with time objects. Does not work for negative years. Requires 4 digits for year.

Usage


fun_date2jd(date_time)

Arguments

date_time
vector containing date and time in FITS format.

Value

vector with Julian Dates.

Author(s)

[Thomas K. Friedli](mailto:thomas.friedli@bluewin.ch)

[Package sunxplrr version 0.1.0 Index]