fun_date_grid {sunxplrr}	R Documentation

Provides a tibble with dates from start to end

Description

Privides a tibble with dates and columns containing year, month and day. The granularity can be chosen from year month or day. Works for all years within the gregorian calendar.

Usage


fun_date_grid(start_date, end_date, granularity = "day")

Arguments

start_date
First date of grid.

end_date
Last date of grid.

granularity
Implemented are "year", "month" and "day".

Value

tibble containing date, and depending on the granularity also year, month and day.

Author(s)

[Thomas K. Friedli](mailto:thomas.friedli@bluewin.ch)

[Package sunxplrr version 0.1.0 Index]