fun_deltaT {sunxplrr}	R Documentation

Calculates delta T in seconds according to formulae by Fred Espenak

Description

Calculates delta T = TD - UT in seconds according to formulae given by Fred Espenak. Before 2005, the polynomial expressions for delta T are from the Five Millennium Canon of Solar Eclipses: -1999 to +3000. Source: http://www.eclipsewise.com/help/deltatpoly2014.html

Usage


fun_deltaT(date_time)

Arguments

date_time
vector containing date and time in FITS format.

Value

value of delta T in seconds.

Author(s)

[Thomas K. Friedli](mailto:thomas.friedli@bluewin.ch)

[Package sunxplrr version 0.1.0 Index]