fun_disc_coordinates {sunxplrr}	R Documentation

Determines heliografic coordinates

Description

Determines heliografic coodinates, including the pixel coordinates from disc center, the angle theta, measured in degrees from the disc center to each disc pixel. For limb pixels theta = 90°. Provided are also the radius r, the angle phi, the heliografic latitude B, the heliografic longitude l and L as the Carrington rotation number for each pixel.

Usage


fun_disc_coordinates(x, hdrlst)

Arguments

x
tibble containing columns with pixel coordinates i and j and additional columns which were all passed to the output tibble z.

hdrlst
list containing image FITS header keywords and values.

Value

z tibble with additional columns to x containing the heliografic coordinate information.

Author(s)

[Thomas K. Friedli](mailto:thomas.friedli@bluewin.ch)

[Package sunxplrr version 0.1.0 Index]