fun_disc_grid {sunxplrr}	R Documentation

Calculates grid of heliografic longitudes and latitudes

Description

Calculates positions of grid pixels for heliografic longitudes and latitudes in intervals of 10 degrees.

Usage


fun_disc_grid(x, hdrlst, grid_each_deg = 10,
  res_each_deg_on_grid = 0.01)

Arguments

x
tibble containing columns with pixel coordinates i and j and additional columns which were all passed to the output tibble z.

hdrlst
list containing image FITS header keywords and values.

grid_each_deg
num number of degrees between grid lines. Default value is one grid line every 10 heliografic degrees.

res_each_deg_on_grid
num indicates the resolution of points of one degree along the grid lines. Default is 0.01, e.g. 100 points per degree.

Value

z tibble with additional column to x containing the flag for grid pixels.

Author(s)

[Thomas K. Friedli](mailto:thomas.friedli@bluewin.ch)

[Package sunxplrr version 0.1.0 Index]