fun_disc_math {sunxplrr}	R Documentation

Implements mathematical image manipulations.

Description

Implements three mathematical image manipulations, including multiplication, substraction and summation.

Usage


fun_disc_math(im1, values_1 = "x", im2, values_2 = "fill",
  method = "mult", values.name = "image")

Arguments

im1
tibble containing at least 3 columns wih pixel coordinates i and j and pixel values of first image.

values_1
name of variable in im1 containing the pixel values.

im2
tibble containing at least 3 columns wih pixel coordinates i and j and pixel values of second image.

values_2
name of variable in im2 containing the pixel values.

method
calculation method. Implemented are thre methods: "mult" for image multiplication, "diff" for image substraction and "summ" for image summation.

values.name
name of variable containing the calculated pixel values.

Value

a tibble with additional column containing the resulting image.

Author(s)

[Thomas K. Friedli](mailto:thomas.friedli@bluewin.ch)

[Package sunxplrr version 0.1.0 Index]