fun_disc_rings {sunxplrr}	R Documentation

Assigns ring number to disc pixels

Description

Computes a table with the ring borders given the number of rings and assigns the number of the ring for each disc pixel.

Usage


fun_disc_rings(x, num.rings = 50)

Arguments

num.rings
number of rings counted from limb to disc center.

Value

z tibble with additional column to x containing the ring number.

Author(s)

[Thomas K. Friedli](mailto:thomas.friedli@bluewin.ch)

[Package sunxplrr version 0.1.0 Index]