fun_extract_col {sunxplrr}	R Documentation

Extracts a single column of image frame

Description

Extracts a single column of image frame.

Usage


fun_extract_col(x, col = NULL)

Arguments

x
tibble containing 3 columns wih pixel coordinates i and j and pixel intensity values x.

row
int index of column to be extracted.

Value

tibble containing 2 columns wih row numbers j and pixel intensity values x.

Author(s)

[Thomas K. Friedli](mailto:thomas.friedli@bluewin.ch)

[Package sunxplrr version 0.1.0 Index]