fun_flop_image {sunxplrr}	R Documentation

Flops the provided image

Description

Flops the provided image by reversing the x-axis of the image (interchanging left and right side). Do not flop for sdo images, which are oriented east to the left and north to the top. The algorithm is from Berry, R. and Burnell, J.: The Handbook of Astronomical Image Processing, 2nd edition 2005, p. 332.

Usage


fun_flop_image(image, hdrlst, header, sdo.image = "FALSE")

Arguments

image
tibble image with i and j pixel coordinates and pixel values x.

hdrlst
list containing image FITS header keywords and values.

header
list containing image FITS header.

sdo.image
boolean switch for dummy use in the case of non sdo calcium images.

Value

tibble with flopped image, updated hdrlst and header.

Author(s)

[Thomas K. Friedli](mailto:thomas.friedli@bluewin.ch)

[Package sunxplrr version 0.1.0 Index]