fun_jd2date {sunxplrr}	R Documentation

Calculates date and time vector from Julian Dates vector

Description

Calculates date and time vector from vector containing Julian Dates objects. Does not work for negative years.

Usage


fun_jd2date(JD)

Value

vector containing date with time in FITS format.

Author(s)

[Thomas K. Friedli](mailto:thomas.friedli@bluewin.ch)

[Package sunxplrr version 0.1.0 Index]