fun_mask_circle {sunxplrr}	R Documentation

Determines a circle given the center coordinates and the radius

Description

Determines limb pixels of a circle with given center coordinates and radius and stores them in a mask. Optionally, additional border pixels may be added to the circle radius.

Usage


fun_mask_circle(x, disc.center, border.pix = 0)

Arguments

x
tibble containing columns with pixel coordinates i and j.

border.pix
number of pixels to be added to the radius of the circle.

disc_center
tibble with center coordinates x_i for image columns and y_j for image rows and radius of the circle.

Value

z tibble with additional column to x containing the limb pixels of the circle.

Author(s)

[Thomas K. Friedli](mailto:thomas.friedli@bluewin.ch)

[Package sunxplrr version 0.1.0 Index]