fun_mask_diff {sunxplrr}	R Documentation

Marks all pixels between two given thresholds

Description

Calculates binarised pixel mask for all pixels above or equal a given lower.threshold but lower than a given upper.threshold.

Usage


fun_mask_diff(x, hdrlst, lower.threshold = 5, upper.threshold = 10,
  method = "relative")

Arguments

x
tibble containing 3 columns wih pixel coordinates i and j and pixel intensity values x.

hdrlst
tibble containing image FITS header keywords and values.

lower.threshold
cutoff threshold in absolute or relative units.

upper.threshold
cutoff threshold in absolute or relative units.

method
If method = 'absolute' then the cutoff thereshold value is in ADU units. Otherwise it is the threshold / 100 fraction of the maximal ADU value given by 2^bitpix.

Value

list containing tibble with 4 columns for i, j pixel coordinates, intensity value x, binarised mask th and tibbles with mask.lower.threshold and mask.upper.threshold values.

Author(s)

[Thomas K. Friedli](mailto:thomas.friedli@bluewin.ch)

[Package sunxplrr version 0.1.0 Index]