fun_mask_fill {sunxplrr}	R Documentation

Fills out a given mask

Description

Fills out a given mask within their biggest contours.

Usage


fun_mask_fill(x, mask = "th")

Arguments

x
tibble containing columns with pixel coordinates i and j and mask.

mask
name of the column in x containing the mask.

Value

z tibble with additional column to x containing the filled mask.

Author(s)

[Thomas K. Friedli](mailto:thomas.friedli@bluewin.ch)

[Package sunxplrr version 0.1.0 Index]