fun_mat2tibbl {sunxplrr}	R Documentation

Umwandlung einer Matrix in ein tibble

Description

Wandelt eine Matrix in ein data.frame oder ein tibble um.

Usage


fun_mat2tibbl(x, tibbl = TRUE)

Arguments

x
matrix with i columns and j rows.

tibbl
boolian If TRUE, the a tibble will be returned, otherwise a data.frame.

Value

a tibble or a data.frame containing 3 columns wih matrix indices i and j and values x.

Author(s)

[Thomas K. Friedli](mailto:thomas.friedli@bluewin.ch)

[Package sunxplrr version 0.1.0 Index]