fun_read_image {sunxplrr}	R Documentation

Reads FITS image from disc

Description

Reads 16 bit or 8 bit FITS image from disc.

Usage


fun_read_image(filename = "sunviewr.fits", maxLines = 5000)

Arguments

maxLines
maximal number of header lines to search for END statement. May be increased for very large headers.

file
output path and file name.

Value

list with image matrix, parsed header and full header vectors.

Author(s)

[Thomas K. Friedli](mailto:thomas.friedli@bluewin.ch)

[Package sunxplrr version 0.1.0 Index]