fun_ring_get {sunxplrr}	R Documentation

Returns ring parameters given a theta value

Description

Computes a table with the ring borders given the number of rings and returns the ring characteristics given a single theta value.

Usage


fun_ring_get(theta = 75, num.rings = 50)

Arguments

theta
angle measured in degrees from the disc center to the limb.

num.rings
number of rings counted from limb to disc center.

Value

table and a vector with ring characteristic for a given theta value.

Author(s)

[Thomas K. Friedli](mailto:thomas.friedli@bluewin.ch)

[Package sunxplrr version 0.1.0 Index]