fun_ring_plt {sunxplrr}	R Documentation

Adds a ring to the image given a theta value

Description

Computes the ring characteristics given a single theta value and adds the inner and outer ring borders to the image.

Usage


fun_ring_plt(x, hdrlst, theta = 75, num.rings = 50)

Arguments

x
tibble containing columns with pixel coordinates i and j, pixel values x from the original image, fill from the disc mask, image from the extracted image disc and theta.

hdrlst
list containing image FITS header keywords and values.

theta
angle measured in degrees from the disc center to the limb.

num.rings
number of rings counted from limb to disc center.

Value

tibble with additional column containing the image with ring borders.

Author(s)

[Thomas K. Friedli](mailto:thomas.friedli@bluewin.ch)

[Package sunxplrr version 0.1.0 Index]