fun_sdo_keywords {sunxplrr}	R Documentation

Updates the header and hdrlst with sdo-specific keywords

Description

Reconstructs the FITS keywords DATE-OBS, NAXISn and WAVELNTH from the provided sdo file name. Updates the header and hdrlst with some other missing keywords.

Usage


fun_sdo_keywords(file.name, header, hdrlst, sdo.image = "FALSE")

Arguments

header
list containing image FITS header.

hdrlst
list containing image FITS header keywords and values.

sdo.image
boolean switch for dummy use in the case of non sdo calcium images.

filename
original file name of sdo image.

Value

list with updated header and hdrlst.

Author(s)

[Thomas K. Friedli](mailto:thomas.friedli@bluewin.ch)

[Package sunxplrr version 0.1.0 Index]