fun_sun_ephem {sunxplrr}	R Documentation

Calculates ephemeris for physical coordinates of the Sun

Description

Calculates ephemeris for physical coordinates of the Sun, including P0, B0, L0, the Sun's apparent diameter in arcsecs and the Carrington rotation number. Formulae from: Meeus,J., Astronomical Algorithms, Willmann-Bell, 1991.

Usage


fun_sun_ephem(hdrlst, sdo.image = "FALSE")

Arguments

hdrlst
list containing image FITS header keywords and values.

sdo.image
boolean switch for case of sdo image.

Value

tibble with P0, B0, L0, the Sun's apparent diameter in arcsecs and the Carrington rotation number.

Author(s)

[Thomas K. Friedli](mailto:thomas.friedli@bluewin.ch)

[Package sunxplrr version 0.1.0 Index]