fun_write_header {sunxplrr}	R Documentation

Writes image header

Description

Writes and closes the image header.

Usage


fun_write_header(x, type = "single", bscale = 1, bzero = 0,
  header = "")

Arguments

x
image data matrix.

type
implemented are "single" and "byte" for 16 bit and 8 bit images.

bscale
pixel_values = bzero + bscale * FITS_value.

bzero
pixel_values = bzero + bscale * FITS_value.

header
string vector with card images of the existing header and of new or modified card images.

Value

z tibble with additional column to x containing the flattened image.

Author(s)

[Thomas K. Friedli](mailto:thomas.friedli@bluewin.ch)

[Package sunxplrr version 0.1.0 Index]