mod_clv_correction {sunxplrr}	R Documentation

Fits the clv function, calibrate it and flattens the image

Description

Fits a six parameter polynom to the limb-variation-function. Additionnally, a plane may be fitted trough the center of the image disc. Removes all fitted non clv variation from the image and flattens it to intensity contrasts. Be aware, that the resulting intensity contrast values should be multiplied with some appropriate factor before converting them back to FITS, since in the FITS image all values are treated as integers.

Usage


mod_clv_correction(x, hdrlst, header, model = "poly_with_plane",
  run = 3, clip.resid.out = "FALSE", sclv.method = "NL",
  sdo.image = "FALSE", light.save = "FALSE")

Arguments

x
tibble containing columns wih pixel coordinates i and j and all the calculated image information as provided by the disc extraction module.

hdrlst
list containing image FITS header keywords and values.

header
list containing image FITS header.

model
implemented are "poly" for polynomial fit with 6 coefficients, "poly_with_plane" for polynomial fit with 6 coefficients and an additional 2 parameter plane through the disc center and a third model named as "poly_with_plane_and nonparametric background" for polynomial fit with 6 coefficients, an additional 2 parameter plane and a nonparametric surface fit in a gam structure.

run
indicates how many iterations for spot and faculae elimination should be run. No more than 3 iterations are implemented.

clip.resid.out
if "TRUE" the clipped residuals from the last fit iteration are returned. In batch mode this output is not needed.

sdo.image
boolean switch for dummy use in the case of non sdo calcium images.

light.save
if TRUE only a small selection of csv files are saved.

Value

tibble containing clv function, calib image and flat image.

Author(s)

[Thomas K. Friedli](mailto:thomas.friedli@bluewin.ch)

[Package sunxplrr version 0.1.0 Index]