mod_disc_image {sunxplrr}	R Documentation

Calculates disc center coordinates and radius and extracts the image

Description

Calculates disc center coordinates and radius and extracts the image. If the original image contains additional text in the image corners, the approximate center coordinates and the approximate radius of the image disc are estimated and the image disk is extracted with an oversized mask. The size of the image and the number of bits per pixel are taken from the FITS header provided by hdrlst.

Usage


mod_disc_image(x, hdrlst, header, threshold = 10, method = "relative",
  cut.threshold = 20, cut.method = "relative", cut.border.pix = 100,
  add.border.pix = 0, image.values.name = "image",
  grid_each_deg = 10, res_each_deg_on_grid = 0.01)

Arguments

x
tibble containing 3 columns wih pixel coordinates i and j and pixel intensity values x.

hdrlst
list containing image FITS header keywords and values.

header
list containing image FITS header.

threshold
cutoff threshold in absolute or relative units.

method
If method = 'absolute' then the cutoff thereshold value is in ADU units. Otherwise it is the threshold / 100 fraction of the maximal ADU value given by 2^bitpix.

cut.threshold
cutoff threshold for images with text in the corners in absolute or relative units.

cut.method
If cut.method = 'absolute' then the cutoff thereshold value for images with text in the corners is in ADU units. Otherwise it is the cut.threshold / 100 fraction of the maximal ADU value given by 2^bitpix.

cut.border.pix
number of pixels to be added to the radius of the circle to extract the image without the text in the corners.

add.border.pix
number of pixels to be added to the radius of the circle to extract the final image.

image.values.name
name of column with image values in final image.

grid_each_deg
num number of degrees between grid lines. Default value is one grid line every 10 heliografic degrees.

res_each_deg_on_grid
num indicates the resolution of points of one degree along the grid lines. Default is 0.01, e.g. 100 points per degree.

Value

list with disc.image, hdrlst and header.

Author(s)

[Thomas K. Friedli](mailto:thomas.friedli@bluewin.ch)

[Package sunxplrr version 0.1.0 Index]