mod_feature_extraction {sunxplrr}	R Documentation

Extracts the plage, enhanced network and active network features

Description

Extracts the plage, enhanced network and active network features by applying cutoff thresholds. Calculates also the hemispheric corrected values for the extracted feature areas and contrasts.

Usage


mod_feature_extraction(x, hdrlst, header, plage.contrast = 1.35,
  en.contrast = 1.25, qn.contrast = 1.1)

Arguments

x
tibble containing 3 columns wih pixel coordinates i and j and flat image pixel values.

hdrlst
list containing image FITS header keywords and values.

header
list containing image FITS header.

plage.contrast
cutoff threshold for plages.

en.contrast
cutoff threshold for enhanced network.

qt.contrast
cutoff threshold for quiet network.

Value

tibble with mask of the extracted features and their contrast values.

Author(s)

[Thomas K. Friedli](mailto:thomas.friedli@bluewin.ch)

[Package sunxplrr version 0.1.0 Index]