mod_fits_import {sunxplrr}	R Documentation

Imports a single FITS file

Description

Imports single FITS file given the file path and name. Returns tibble with image, tibble with hdrlst and character vector header.

Usage


mod_fits_import(inp_data_path, inp_file_name, sdo.image = "FALSE",
  cut.image = "FALSE")

Arguments

inp_data_path
string with input data path.

inp_file_name
string with input file name.

sdo.image
if TRUE the image is originally an imported sdo jpg-file.

cut.image
if TRUE the image contains additional text in the corners.

Value

list with fitsim, hdrlst and header.

Author(s)

[Thomas K. Friedli](mailto:thomas.friedli@bluewin.ch)

[Package sunxplrr version 0.1.0 Index]