mod_index_calculation {sunxplrr}	R Documentation

Calculates activity indices and provides data for images and charts

Description

Calculates activity indices and provides input data for FITS images and charts.

Usage


mod_index_calculation(x, hdrlst, header)

Arguments

x
tibble containing the previously calculated images.

hdrlst
list containing image FITS header keywords and values.

header
list containing image FITS header.

Value

list with images, hdrlst, header, indices, charts and syopsis.

Author(s)

[Thomas K. Friedli](mailto:thomas.friedli@bluewin.ch)

[Package sunxplrr version 0.1.0 Index]