mod_output {sunxplrr}	R Documentation

Saves results to disk

Description

Saves results to disk.

Usage


mod_output(images, hdrlst, header, total.indices, hemisphere.indices,
  latitude.indices, chart.indices, charts, synopsis.indices, synopsis,
  inp_file_name, rds.output = "FALSE", full.output = "FALSE",
  light.save = "FALSE", fits.save = "FALSE", jpg.save = "FALSE",
  out_data_path)

Arguments

images
tibble containing the previously calculated images.

hdrlst
list containing image FITS header keywords and values.

header
list containing image FITS header.

rds.output
if TRUE all results are saved as R data file.

full.output
if TRUE the full image data table is saved as csv file.

light.save
if TRUE only a small selection of csv files are saved.

out_data_path
full path to output directory.

fits.output
if TRUE some fits images are saved.

jpg.output
if TRUE some daily charts are saved as jpg files.

Value

list containing images, hdrlst, header, indices and charts.

Author(s)

[Thomas K. Friedli](mailto:thomas.friedli@bluewin.ch)

[Package sunxplrr version 0.1.0 Index]