wrap_mod_calcium {sunxplrr}	R Documentation

Runs the sunviewr modules for the calcium image analysis

Description

Runs the sunviewr modules for the exploration of calcium images and the determination of activity indices for one image.

Usage


wrap_mod_calcium(inp_file_name, sdo.image, inp_data_path, out_data_path,
  rds.output = "FALSE", full.output = "FALSE", light.save = "FALSE",
  fits.save = "FALSE", jpg.save = "FALSE")

Arguments

inp_file_name
input file name with extension.

sdo.image
if TRUE, param.lst for SDO image analysis is chosen.

inp_data_path
full path to input directory.

out_data_path
full path to output directory.

rds.output
if TRUE all results are saved as R data file.

full.output
if TRUE the full image data table is saved as csv file.

light.save
if TRUE only a small selection of csv files are saved.

fits.output
if TRUE some fits images are saved.

jpg.output
if TRUE some daily charts are saved as jpg files.

Value

list with resulting image and header.

Author(s)

[Thomas K. Friedli](mailto:thomas.friedli@bluewin.ch)

[Package sunxplrr version 0.1.0 Index]